refactor: change type for monthCount

+++ account-processing-ms/src/main/java/ru/t1/nour/microservice/model/dto/kafka/ClientProductEventDTO.java
@@ -15,6 +15,6 @@ public class ClientProductEventDTO {
     private String productKey;
     private String status;
     private BigDecimal requestedAmount;
-    private Long monthCount;
+    private Integer monthCount;
     private String eventType;
 }
