+++ client-processing-ms/src/main/java/ru/t1/nour/microservice/model/dto/request/ClientProductCreateRequest.java
@@ -19,5 +19,5 @@ public class ClientProductCreateRequest {
     BigDecimal requestedAmount;
 
     @Min(1)
-    Long monthCount;
+    Integer monthCount;
 }
