+++ credit-processing-ms/src/main/java/ru/t1/nour/microservice/model/dto/kafka/ClientProductEventDTO.java
@@ -18,6 +18,6 @@ public class ClientProductEventDTO {
     private String productKey;
     private String status;
     private BigDecimal requestedAmount;
-    private Long monthCount;
+    private Integer monthCount;
     private String eventType;
 }
